* 'develop' of github.com:BlueprintSys/blueprint: (67 commits)
  Add swagger documentation
  add fixes for pull request comments
  #17705 add method for /sessions/select and test for AccessControl
  modify method for getting locked/active license info
  fix according to comments for pull request
  fix comments for pull request
  adds Basic Correctness code analysis
  Update development setup
  Fixed Storyteller based on pull request comments
  CreateUser now updates UserId for newly created user, add CreateSession which takes IUser as argument, modify AccessControl tests to use CreateSession
  Revert "Merge pull request #223 from BlueprintSys/db"
  #177049 add test for licenses/active, licenses/locked and licenses/transactions?days=1&consumerType=
  adds documentation for LicensesController
  Initial Storyteller tests implementation including models and an unimplemented test case
  Create a SQL Job that will perform database maintenancre activities
  #177046 add wrapper for GET /licenses/transactions and test
  fixes SQL paramter captialization
  Changed all GetStatus() functions in AccessControl, AdminStore & FileStore model classes to take an optional list of expected status codes.
  fixes EndSession
  check the current end time before expiring a session
  ...

+++ blueprint-automationframework/Test-Doubles/AccessControlDouble/AccessControlDouble/Controllers/SessionsController.cs
@@ -1,7 +1,5 @@
 ﻿using System;
-using System.Linq;
 using System.Net.Http;
-using System.Net.Http.Headers;
 using System.Threading.Tasks;
 using System.Web.Http;
 using System.Web.Http.Description;
@@ -24,6 +22,31 @@ namespace AccessControlDouble.Controllers
             return WebUtils.CreateUri(Request.RequestUri, WebApiConfig.AccessControl, WebApiConfig.SVC_PATH);
         }
 
+        /// <summary>
+        /// Writes a line into the log file.
+        /// </summary>
+        /// <param name="line">The line to write.</param>
+        private static void WriteLine(string line)
+        {
+            using (LogFile logFile = new LogFile(WebApiConfig.LogFile))
+            {
+                logFile.WriteLine(line);
+            }
+        }
+
+        /// <summary>
+        /// Writes a formatted line into the log file.
+        /// </summary>
+        /// <param name="format">The format string to write.</param>
+        /// <param name="args">The format arguments.</param>
+        private static void WriteLine(string format, params Object[] args)
+        {
+            using (LogFile logFile = new LogFile(WebApiConfig.LogFile))
+            {
+                logFile.WriteLine(format, args);
+            }
+        }
+
         #endregion Private functions
 
         /// <summary>
@@ -39,17 +62,35 @@ namespace AccessControlDouble.Controllers
         [ResponseType(typeof(HttpResponseMessage))]
         public async Task<IHttpActionResult> Get(int uid)
         {
-            using (LogFile logFile = new LogFile(WebApiConfig.LogFile))
+            string thisNamespace = nameof(AccessControlDouble);
+            string thisClassName = nameof(SessionsController);
+            string thisMethodName = nameof(Get);
+
             using (HttpClient http = new HttpClient())
             {
-                logFile.WriteLine("Called AccessControlDouble.SessionsController.Get({0})", uid);
+                await Task.Run(() =>
+                {
+                    WriteLine("Called {0}.{1}.{2}({3})", thisNamespace, thisClassName, thisMethodName, uid);
+                });
+
+                // If the test wants to inject a custom status code, return that instead of the real value.
+                if (WebApiConfig.StatusCodeToReturn["GET"].HasValue)
+                {
+                    return ResponseMessage(Request.CreateResponse(WebApiConfig.StatusCodeToReturn["GET"].Value));
+                }
 
                 WebUtils.ConfigureHttpClient(http, Request, WebApiConfig.AccessControl);
                 var uri = CreateUri();
 
-                logFile.WriteLine("Calling http.GetAsync({0})", uid);
+                await Task.Run(() =>
+                {
+                    WriteLine("Calling http.GetAsync({0})", uid);
+                });
                 var result = await http.GetAsync(uri);
-                WebUtils.LogRestResponse(logFile, result);
+                await Task.Run(() =>
+                {
+                    WebUtils.LogRestResponse(WebApiConfig.LogFile, result);
+                });
 
                 return ResponseMessage(result);
             }
@@ -68,17 +109,35 @@ namespace AccessControlDouble.Controllers
         [ResponseType(typeof(HttpResponseMessage))]
         public async Task<IHttpActionResult> Get(string ps = "100", string pn = "1")
         {
-            using (LogFile logFile = new LogFile(WebApiConfig.LogFile))
+            string thisNamespace = nameof(AccessControlDouble);
+            string thisClassName = nameof(SessionsController);
+            string thisMethodName = nameof(Get);
+
             using (HttpClient http = new HttpClient())
             {
-                logFile.WriteLine("Called AccessControlDouble.SessionsController.Get('{0}', '{1}')", ps, pn);
+                await Task.Run(() =>
+                {
+                    WriteLine("Called {0}.{1}.{2}('{3}', '{4}')", thisNamespace, thisClassName, thisMethodName, ps, pn);
+                });
+
+                // If the test wants to inject a custom status code, return that instead of the real value.
+                if (WebApiConfig.StatusCodeToReturn["GET"].HasValue)
+                {
+                    return ResponseMessage(Request.CreateResponse(WebApiConfig.StatusCodeToReturn["GET"].Value));
+                }
 
                 WebUtils.ConfigureHttpClient(http, Request, WebApiConfig.AccessControl);
                 var uri = CreateUri();
 
-                logFile.WriteLine("Calling http.GetAsync('{0}', '{1}')", ps, pn);
+                await Task.Run(() =>
+                {
+                    WriteLine("Calling http.GetAsync('{0}', '{1}')", ps, pn);
+                });
                 var result = await http.GetAsync(uri);
-                WebUtils.LogRestResponse(logFile, result);
+                await Task.Run(() =>
+                {
+                    WebUtils.LogRestResponse(WebApiConfig.LogFile, result);
+                });
 
                 return ResponseMessage(result);
             }
@@ -98,18 +157,36 @@ namespace AccessControlDouble.Controllers
         [ResponseType(typeof(HttpResponseMessage))]
         public async Task<IHttpActionResult> Post(int uid, string userName, int licenseLevel, bool isSso = false)
         {
-            using (LogFile logFile = new LogFile(WebApiConfig.LogFile))
+            string thisNamespace = nameof(AccessControlDouble);
+            string thisClassName = nameof(SessionsController);
+            string thisMethodName = nameof(Post);
+
             using (HttpClient http = new HttpClient())
             {
-                logFile.WriteLine("Called AccessControlDouble.SessionsController.Post({0}, '{1}', {2}, {3})",
-                    uid, userName, licenseLevel, isSso);
+                await Task.Run(() =>
+                {
+                    WriteLine("Called {0}.{1}.{2}({3}, '{4}', {5}, {6})",
+                        thisNamespace, thisClassName, thisMethodName, uid, userName, licenseLevel, isSso);
+                });
+
+                // If the test wants to inject a custom status code, return that instead of the real value.
+                if (WebApiConfig.StatusCodeToReturn["POST"].HasValue)
+                {
+                    return ResponseMessage(Request.CreateResponse(WebApiConfig.StatusCodeToReturn["POST"].Value));
+                }
 
                 WebUtils.ConfigureHttpClient(http, Request, WebApiConfig.AccessControl);
                 var uri = CreateUri();
 
-                logFile.WriteLine("Calling http.PostAsJsonAsync('{0}', {1})", uri.ToString(), uid);
+                await Task.Run(() =>
+                {
+                    WriteLine("Calling http.PostAsJsonAsync('{0}', {1})", uri.ToString(), uid);
+                });
                 var result = await http.PostAsJsonAsync(uri, uid);
-                WebUtils.LogRestResponse(logFile, result);
+                await Task.Run(() =>
+                {
+                    WebUtils.LogRestResponse(WebApiConfig.LogFile, result);
+                });
 
                 return ResponseMessage(result);
             }
@@ -129,17 +206,35 @@ namespace AccessControlDouble.Controllers
         [ResponseType(typeof(HttpResponseMessage))]
         public async Task<IHttpActionResult> Put(string op = "", int aid = 0)
         {
-            using (LogFile logFile = new LogFile(WebApiConfig.LogFile))
+            string thisNamespace = nameof(AccessControlDouble);
+            string thisClassName = nameof(SessionsController);
+            string thisMethodName = nameof(Put);
+
             using (HttpClient http = new HttpClient())
             {
-                logFile.WriteLine("Called AccessControlDouble.SessionsController.Put('{0}', {1})", op, aid);
+                await Task.Run(() =>
+                {
+                    WriteLine("Called {0}.{1}.{2}('{3}', {4})", thisNamespace, thisClassName, thisMethodName, op, aid);
+                });
+
+                // If the test wants to inject a custom status code, return that instead of the real value.
+                if (WebApiConfig.StatusCodeToReturn["PUT"].HasValue)
+                {
+                    return ResponseMessage(Request.CreateResponse(WebApiConfig.StatusCodeToReturn["PUT"].Value));
+                }
 
                 WebUtils.ConfigureHttpClient(http, Request, WebApiConfig.AccessControl);
                 var uri = CreateUri();
 
-                logFile.WriteLine("Calling http.PutAsync('{0}', {1})", op, aid);
+                await Task.Run(() =>
+                {
+                    WriteLine("Calling http.PutAsync('{0}', {1})", op, aid);
+                });
                 var result = await http.PutAsync(uri, Request.Content);
-                WebUtils.LogRestResponse(logFile, result);
+                await Task.Run(() =>
+                {
+                    WebUtils.LogRestResponse(WebApiConfig.LogFile, result);
+                });
 
                 return ResponseMessage(result);
             }
@@ -155,17 +250,35 @@ namespace AccessControlDouble.Controllers
         [ResponseType(typeof(HttpResponseMessage))]
         public async Task<IHttpActionResult> Delete()
         {
-            using (LogFile logFile = new LogFile(WebApiConfig.LogFile))
+            string thisNamespace = nameof(AccessControlDouble);
+            string thisClassName = nameof(SessionsController);
+            string thisMethodName = nameof(Delete);
+
             using (HttpClient http = new HttpClient())
             {
-                logFile.WriteLine("Called AccessControlDouble.SessionsController.Delete()");
+                await Task.Run(() =>
+                {
+                    WriteLine("Called {0}.{1}.{2}()", thisNamespace, thisClassName, thisMethodName);
+                });
+
+                // If the test wants to inject a custom status code, return that instead of the real value.
+                if (WebApiConfig.StatusCodeToReturn["DELETE"].HasValue)
+                {
+                    return ResponseMessage(Request.CreateResponse(WebApiConfig.StatusCodeToReturn["DELETE"].Value));
+                }
 
                 WebUtils.ConfigureHttpClient(http, Request, WebApiConfig.AccessControl);
                 var uri = CreateUri();
 
-                logFile.WriteLine("Calling http.DeleteAsync()");
+                await Task.Run(() =>
+                {
+                    WriteLine("Calling http.DeleteAsync()");
+                });
                 var result = await http.DeleteAsync(uri);
-                WebUtils.LogRestResponse(logFile, result);
+                await Task.Run(() =>
+                {
+                    WebUtils.LogRestResponse(WebApiConfig.LogFile, result);
+                });
 
                 return ResponseMessage(result);
             }
